[sailfish-iptables-api-plugin] Implemented new design of iptables API.
Allowing/denying incoming/outcoming connections with ip, network (mask converted
to cidr notation), port, port range and service combinations are supported. A
separate method for each action. Each succesful method call results in appropriate
signal.

Messages take protocol and operation (add|remove) as parameter.

Operation defaults to add. Operation is not case sensitive.

Protocol is not mandatory if it can be retrieved with getservbyname or
getservbyport. This usually defaults to tcp according to the services
list. Protocol is not case sensitive.

Service name is not case sensitive. Policy is not case sensitive.